fix(protective_stop): widen rebond watchdog to uint64_t, fix sock init on early return
hb_ms is adopted unclamped from the wire, so hb_ms * max_missed in a 32-bit
accumulator could wrap and produce a watchdog threshold below the machine's
own bond-drop timeout -- inverting the invariant pstop_session_rebond_after_ms
exists to guarantee. Widen to uint64_t end-to-end, matching upstream's
sess_rebond_after_ms().

pstop_session_open() also returned -EINVAL before ever assigning s->sock, so
a zero-initialised session left sock == 0 (stdin) on that path and a later
close would target it. Assign sock = -1 before parameter validation so the
"sock is -1 whenever not open" invariant holds on every exit.

+++ applications/protective_stop/src/session.h
@@ -94,9 +94,12 @@ void pstop_session_close(struct pstop_session *s);
 uint32_t pstop_session_send_period_ms(const struct pstop_session *s);
 
 /* How long without a reply before this session re-bonds. Always greater than
- * the machine's own hb_ms * max_missed bond-drop timeout.
+ * the machine's own hb_ms * max_missed bond-drop timeout. hb_ms is adopted
+ * unclamped from the wire, so this widens to uint64_t rather than clamping it
+ * -- a 32-bit accumulator would wrap and could undercut the machine's own
+ * timeout.
  */
-uint32_t pstop_session_rebond_after_ms(const struct pstop_session *s);
+uint64_t pstop_session_rebond_after_ms(const struct pstop_session *s);
 
 /* Is this session due to transmit on this tick? */
 bool pstop_session_due(const struct pstop_session *s, uint64_t now_ms);
